+++ docs/changelog.md
@@ -63,6 +63,10 @@
 
 ## 更新说明详情
 
+1.10.57 fix: fix stock_zh_valuation_baidu interface
+
+    1. 修复 stock_zh_valuation_baidu 接口
+
 1.10.56 fix: fix stock_zh_b_daily interface
 
     1. 修复 stock_zh_b_daily 接口
@@ -2637,6 +2641,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.10.57 fix: fix stock_zh_valuation_baidu interface
 
 1.10.56 fix: fix stock_zh_b_daily interface